requirements.txt
# ============================================================
#  requirements.txt — Pipeline ETL Industrialisé
# ============================================================

# ─── Extraction ──────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0               # Parser HTML rapide pour BeautifulSoup

# ─── Données ─────────────────────────────────────────────────
pandas==2.1.4
numpy==1.26.4

# ─── Base de données ─────────────────────────────────────────
sqlalchemy==2.0.25
psycopg2-binary==2.9.9    # Connecteur PostgreSQL (Supabase)

# ─── Spark ───────────────────────────────────────────────────
pyspark==3.5.1

# ─── MinIO / Stockage ────────────────────────────────────────
boto3==1.34.34             # Client S3 pour MinIO
minio==7.2.3               # SDK MinIO optionnel

# ─── Configuration ───────────────────────────────────────────
python-dotenv==1.0.1       # Chargement du .env
pyyaml==6.0.1              # Lecture de config.yaml

# ─── Airflow (installation séparée recommandée) ──────────────
# apache-airflow==2.8.1
# apache-airflow-providers-amazon==8.18.0

# ─── Tests ───────────────────────────────────────────────────
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
